remove SurfStrategy trait, only has one impl anyway
